[Core] Use the test case duration in the JUnit Formatter.
Let the JUnit Formatter use the duration of the test case result
instead of summing up the durations of the test step results.

Also refactor the TimeService to have the methods startTime(),
currentTime(), and finishTime(). These methods actually do the same
thing but having them all enables/simplifies the mocking of time in
tests.